+++ manage.py
@@ -1,4 +1,4 @@
-from app import server, db
+from app import server, db, Filter
 
 
 def create_all():
@@ -6,5 +6,12 @@ def create_all():
         db.create_all()
 
 
-if __name__ == "__main__":
-    create_all()
+def delete_filter(filter_name):
+    with server.app_context():
+        filter_ = Filter.query.filter_by(name=filter_name).first()
+        if filter_:
+            db.session.delete(filter_)
+            db.session.commit()
+            print(f"Successfully deleted filter '{filter_name}'.")
+        else:
+            print(f"No filter '{filter_name}' found.")